backend/requirements.txt
asgiref==3.7.2
colorama==0.4.6
coverage==7.3.2
cssselect==1.2.0
distlib==0.3.7
Django==5.0.1
django-cors-headers==4.3.1
django-seed==0.3.1
django-widget-tweaks==1.5.0
django-with-asserts==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==19.11.0
filelock==3.12.4
halo==0.0.31
iniconfig==2.0.0
Levenshtein==0.25.0
libgravatar==1.0.4
log-symbols==0.0.14
lxml==4.9.3
packaging==23.1
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.2
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3.post1
rapidfuzz==3.6.1
sendgrid==6.11.0
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
termcolor==2.4.0
toposort==1.10
typing_extensions==4.8.0
virtualenv==20.24.5